Update README.md
As discussed with  @ccarouge @aidanheerdegen this afternoon, this text is currently misleading as a mixed approach is being used where eventually the plan is to move to a per repository access basis. Perhaps text like I've suggested could be used eventually? 

Alternatively, can this text please be updated to reflect the current suggested approach? Have admins of all the current organisation's repos been informed of the change / update of the READMEs as appropriate?